Replace redis with aioredis
This affects a lot of files because it turned a lot of widely-used methods red.

Note: I had to remove Parrot's ability to wait for the database to finish loading if it is still doing so while Parrot is starting up. This happens in the Parrot.run() method, but this is method is called in blue code, so I couldn't let it turn red like it would need to be for it to continue to work with aioredis's async ping command, so... I lost my train of thought. Just make sure the database is loaded before you start Parrot.