Create method for getting corpus with IDs
Messages are stored in as "message ID: message content" key-value pairs in the database. Parrot never uses the message IDs, so I just didn't fetch them. However, having the IDs is useful for certain debugging and maintenance tasks, so I made this new method that preserves the message IDs within a corpus.